WFLY-4294 remove RUNTIME stage for :enable :disable operation for (xa)datasource
