+++ connector/src/main/java/org/jboss/as/connector/subsystems/datasources/DataSourceDisable.java
@@ -22,27 +22,11 @@
 
 package org.jboss.as.connector.subsystems.datasources;
 
-import static org.jboss.as.connector.subsystems.datasources.Constants.JNDI_NAME;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ENABLED;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
-
-import javax.sql.DataSource;
-import java.util.List;
-
 import org.jboss.as.connector.logging.ConnectorLogger;
-import org.jboss.as.connector.services.datasources.statistics.DataSourceStatisticsService;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.OperationStepHandler;
-import org.jboss.as.controller.PathAddress;
-import org.jboss.as.controller.PathElement;
-import org.jboss.as.controller.registry.ManagementResourceRegistration;
-import org.jboss.as.controller.registry.Resource;
-import org.jboss.as.naming.deployment.ContextNames;
 import org.jboss.dmr.ModelNode;
-import org.jboss.msc.service.ServiceController;
-import org.jboss.msc.service.ServiceName;
-import org.jboss.msc.service.ServiceRegistry;
 
 /**
  * Operation handler responsible for disabling an existing data-source.
@@ -50,151 +34,17 @@ import org.jboss.msc.service.ServiceRegistry;
  * @author John Bailey
  */
 public class DataSourceDisable implements OperationStepHandler {
-    static final DataSourceDisable LOCAL_INSTANCE = new DataSourceDisable(false);
-    static final DataSourceDisable XA_INSTANCE = new DataSourceDisable(true);
+    static final DataSourceDisable INSTANCE = new DataSourceDisable();
 
-    private final boolean xa;
-
-    public DataSourceDisable(boolean xa) {
+    public DataSourceDisable() {
         super();
-        this.xa = xa;
     }
 
-    public void execute(OperationContext context, ModelNode operation) {
-
-        final ManagementResourceRegistration datasourceRegistration = context.getResourceRegistrationForUpdate();
-        final Resource resource = context.readResourceForUpdate(PathAddress.EMPTY_ADDRESS);
-        final ModelNode model = resource.getModel();
-        model.get(ENABLED).set(false);
+    public void execute(OperationContext context, ModelNode operation) throws OperationFailedException {
 
         if (context.isNormalServer()) {
-
-            DataSourceStatisticsService.removeStatisticsResources(resource);
-
-            if (context.isResourceServiceRestartAllowed()) {
-                context.addStep(new OperationStepHandler() {
-                    public void execute(final OperationContext context, ModelNode operation) throws OperationFailedException {
-
-                        final ModelNode address = operation.require(OP_ADDR);
-                        final String dsName = PathAddress.pathAddress(address).getLastElement().getValue();
-                        final String jndiName = JNDI_NAME.resolveModelAttribute(context, model).asString();
-                        final ContextNames.BindInfo bindInfo = ContextNames.bindInfoFor(jndiName);
-
-                        final ServiceRegistry registry = context.getServiceRegistry(true);
-
-                        final ServiceName dataSourceServiceName = context.getCapabilityServiceName(Capabilities.DATA_SOURCE_CAPABILITY_NAME, dsName, DataSource.class);
-                        final ServiceController<?> dataSourceController = registry.getService(dataSourceServiceName);
-                        if (dataSourceController != null) {
-                            if (ServiceController.State.UP.equals(dataSourceController.getState())) {
-                                dataSourceController.setMode(ServiceController.Mode.NEVER);
-                            } else {
-                                throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceNotEnabled("Data-source", dsName));
-                            }
-                        } else {
-                            throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceNotAvailable("Data-source", dsName));
-                        }
-                        context.removeService(CommonDeploymentService.getServiceName(bindInfo));
-                        context.removeService(dataSourceServiceName.append(Constants.STATISTICS));
-                        final ServiceName referenceServiceName = DataSourceReferenceFactoryService.SERVICE_NAME_BASE.append(dsName);
-                        final ServiceController<?> referenceController = registry.getService(referenceServiceName);
-                        if (referenceController != null) {
-                            context.removeService(referenceController);
-                        }
-
-                        final ServiceName binderServiceName = bindInfo.getBinderServiceName();
-
-                        final ServiceController<?> binderController = registry.getService(binderServiceName);
-                        if (binderController != null) {
-                            context.removeService(binderController);
-                        }
-
-                        final ServiceName dataSourceConfigServiceName = DataSourceConfigService.SERVICE_NAME_BASE.append(dsName);
-                        final ServiceController<?> dataSourceConfigController = registry.getService(dataSourceConfigServiceName);
-
-
-                        final List<ServiceName> serviceNames = registry.getServiceNames();
-
-
-                        final ServiceName xaDataSourceConfigServiceName = XADataSourceConfigService.SERVICE_NAME_BASE.append(dsName);
-                        final ServiceController<?> xaDataSourceConfigController = registry.getService(xaDataSourceConfigServiceName);
-
-
-                        for (ServiceName name : serviceNames) {
-                            if (dataSourceConfigServiceName.append("connection-properties").isParentOf(name)) {
-                                final ServiceController<?> connPropertyController = registry.getService(name);
-
-                                if (connPropertyController != null) {
-                                    if (ServiceController.State.UP.equals(connPropertyController.getState())) {
-                                        connPropertyController.setMode(ServiceController.Mode.NEVER);
-                                    } else {
-                                        throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceAlreadyStarted("Data-source.connectionProperty", name));
-                                    }
-                                } else {
-                                    throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceNotAvailable("Data-source.connectionProperty", name));
-                                }
-                            }
-                            if (xaDataSourceConfigServiceName.append("xa-datasource-properties").isParentOf(name)) {
-                                final ServiceController<?> xaConfigPropertyController = registry.getService(name);
-
-                                if (xaConfigPropertyController != null) {
-                                    if (ServiceController.State.UP.equals(xaConfigPropertyController.getState())) {
-                                        xaConfigPropertyController.setMode(ServiceController.Mode.NEVER);
-                                    } else {
-                                        throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceAlreadyStarted("Data-source.xa-config-property", name));
-                                    }
-                                } else {
-                                    throw new OperationFailedException(ConnectorLogger.ROOT_LOGGER.serviceNotAvailable("Data-source.xa-config-property", name));
-                                }
-                            }
-                        }
-
-
-                        if (xaDataSourceConfigController != null) {
-                            context.removeService(xaDataSourceConfigController);
-                        }
-
-                        if (dataSourceConfigController != null) {
-                            context.removeService(dataSourceConfigController);
-                        }
-
-                        context.completeStep(new OperationContext.RollbackHandler() {
-                            @Override
-                            public void handleRollback(OperationContext context, ModelNode operation) {
-                                try {
-                                    reEnable(context, operation, datasourceRegistration);
-                                } catch (OperationFailedException e) {
-                                    // ignored
-                                }
-                            }
-                        });
-                    }
-                }, OperationContext.Stage.RUNTIME);
-            } else {
-                context.addStep(new OperationStepHandler() {
-                    @Override
-                    public void execute(OperationContext context, ModelNode operation) throws OperationFailedException {
-                        context.reloadRequired();
-                        context.completeStep(OperationContext.RollbackHandler.REVERT_RELOAD_REQUIRED_ROLLBACK_HANDLER);
-                    }
-                }, OperationContext.Stage.RUNTIME);
-            }
+            throw ConnectorLogger.ROOT_LOGGER.legacyOperation();
         }
-        context.stepCompleted();
-    }
 
-    private void reEnable(final OperationContext context, final ModelNode operation, final ManagementResourceRegistration datasourceRegistration) throws OperationFailedException {
-        if (context.isNormalServer()) {
-            PathAddress addr = PathAddress.pathAddress(operation.get(OP_ADDR));
-            Resource resource = context.getOriginalRootResource();
-            for (PathElement element : addr) {
-                resource = resource.getChild(element);
-            }
-            DataSourceEnable.addServices(context, operation, datasourceRegistration,
-                    Resource.Tools.readModel(resource), isXa());
-        }
-    }
-
-    public boolean isXa() {
-        return xa;
     }
 }
