+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jca/metrics/data-sources/basic-attributes.xml
@@ -1,7 +1,7 @@
 <subsystem xmlns="urn:jboss:domain:datasources:3.0">
             <datasources>
                 <datasource jndi-name="java:jboss/datasources/DS"
-                            pool-name="DS" enabled="true">
+                            pool-name="DS" enabled="true" statistics-enabled="true">
                     <connection-url>jdbc:h2:mem:test;DB_CLOSE_DELAY=-1</connection-url>
                     <driver>h2</driver>
                 </datasource>
