+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jca/statistics/DataSourcePoolStatisticsTestCase.java
@@ -95,11 +95,13 @@ public class DataSourcePoolStatisticsTestCase extends JcaStatisticsBase {
         }
 
         operation = new ModelNode();
-        operation.get(OP).set("enable");
+        operation.get(OP).set("write-attribute");
+        operation.get("name").set("enabled");
+        operation.get("value").set(true);
         operation.get(OP_ADDR).set(address);
 
         executeOperation(operation);
-
+        reload();
         return address;
     }
